pyproject.toml
[project]
name = "prj-arxfind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.12",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
]
